@@ -5,7 +5,7 @@ import com.smartbear.ready.plugin.postman.utils.PostmanJsonUtil;
 import net.sf.json.JSON;
 import net.sf.json.JSONObject;
 import org.apache.commons.io.FileUtils;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 
 import java.io.File;
@@ -13,6 +13,7 @@ import java.nio.charset.StandardCharsets;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.CoreMatchers.instanceOf;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class PostmanCollectionFactoryTest {
     private final String COLLECTION_V1_PATH = "/Postman_Echo.postman_collection";
@@ -20,9 +21,11 @@ public class PostmanCollectionFactoryTest {
     private final String COLLECTION_V2_1PATH = "/Postman_Echo.postman_collection_v2.1";
     private final String COLLECTION_FROM_URL_PATH = "/Postman_Echo_from_url.postman_collection_v2.1";
 
-    @Test(expected = PostmanCollectionUnsupportedVersionException.class)
-    public void postmanCollectionV1ImportThrowsUnsupportedVersionException() throws Exception {
-        getCollectionFromFile(getClass().getResource(COLLECTION_V1_PATH).getPath());
+    @Test
+    public void postmanCollectionV1ImportThrowsUnsupportedVersionException() {
+        assertThrows(PostmanCollectionUnsupportedVersionException.class, () -> {
+            getCollectionFromFile(getClass().getResource(COLLECTION_V1_PATH).getPath());
+        });
     }
 
     @Test
